add work with recources files

+++ Task5 StringNumber/Model/ConverterToText.cs
@@ -8,16 +8,16 @@ namespace Task5_StringNumber.Model
 {
     public abstract class ConverterToText
     {
-        protected Dictionary<int, string> _first100;
-        protected Dictionary<int, string> _first100FemaleChanges;
-        protected Dictionary<int, string> _hundreds;
-        protected Dictionary<int, string> _multiplesOf1000;
-        protected Dictionary<int, string> _multiplesOf1000Form234;
-        protected Dictionary<int, string> _multiplesOf1000Form5;
+        protected Dictionary<long, string> _first100;
+        protected Dictionary<long, string> _first100FemaleChanges;
+        protected Dictionary<long, string> _hundreds;
+        protected Dictionary<long, string> _multiplesOf1000;
+        protected Dictionary<long, string> _multiplesOf1000Form234;
+        protected Dictionary<long, string> _multiplesOf1000Form5;
         protected string _negative;
         protected int _maxRank;
 
         protected abstract void LoadResources();
-        public abstract string Convert(int value);
+        public abstract string Convert(long value);
     }
 }
