+++ Task5 StringNumber/Model/ValidationInboxParameters/InboxParameters.cs
@@ -3,12 +3,13 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using Task5_StringNumber.Resources;
 
 namespace Task5_StringNumber.Model.ValidationInboxParameters
 {
     public struct InboxParameters
     {
-        public int Number { get; set; }
+        public long Number { get; set; }
         public Local Region { get; set; }
     }
 }
